src/tests/e2e/validation/requirements.txt
# Copyright (C) 2024-2026 Intel Corporation
# SPDX-License-Identifier: Apache-2.0

aiohttp==3.13.3
pytest==7.4.4 # MIT
allure-pytest==2.13.5 # Apache 2.0
requests==2.32.4 # Apache 2.0
kr8s==0.17.4 # BSD 3-Clause License
sniffio==1.3.1 # MIT
pytest-timeout==2.3.1 # MIT
sentence-transformers==5.1.2 # Apache 2.0
rouge==1.0.1 # Apache 2.0
sherpa-onnx==1.12.23 # Apache 2.0
pydub==0.25.1 # MIT
